requirements.txt
paramiko==2.4.2
PyNaCl==1.3.0
termcolor==1.1.0
scp==0.13.0
nacl==0.0.0
PyYAML==3.13
